Added GetRealtorByHomeId fx to Put / Delete routes (so only the realtor who created these homes can modify). This is done by finding the person who is requesting the update and comparing against the one who created these resources via POST (which was saved in db)

+++ src/home/home.controller.ts
@@ -8,6 +8,7 @@ import {
   Query,
   ParseIntPipe,
   Body,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { HomeService } from './home.service';
 import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './dto/home.dto';
@@ -58,16 +59,28 @@ export class HomeController {
   }
 
   @Put(':id')
-  updateHome(
+  async updateHome(
     @Param('id', ParseIntPipe) id: number,
     @Body() body: UpdateHomeDto,
+    @User() user: UserInfo,
   ): Promise<HomeResponseDto> {
     console.log('this is from controller', id, body);
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException();
+    } // left part fetched from db vs right part is from the decorator (interceptor adds this from request)
     return this.homeService.updateHomeById(id, body);
   }
 
   @Delete(':id')
-  deleteHome(@Param('id', ParseIntPipe) id: number) {
+  async deleteHome(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInfo,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException();
+    } // left part fetched from db vs right part is from the decorator (interceptor adds this from request)
     return this.homeService.deleteHome(id);
   }
 }
